Hoist app fixture across experiemental API tests
